Functional test adding turma alunos

+++ escola/views_materia.py
@@ -1,23 +1,27 @@
 #  Developed by Vinicius José Fritzen
-#  Last Modified 12/04/19 13:19.
+#  Last Modified 25/04/19 22:53.
 #  Copyright (c) 2019  Vinicius José Fritzen and Albert Angel Lanzarini
+import logging
 
 from django.contrib.auth.mixins import LoginRequiredMixin
+from django.contrib.auth.views import redirect_to_login
 from django.http import HttpResponseRedirect
 from django.shortcuts import get_object_or_404, render
 from django.urls import reverse
 from django.views.generic import DetailView
-from guardian.decorators import permission_required as permission_required_obj
+from rolepermissions.checkers import has_object_permission
 
 from escola.decorators import is_user_escola
 from escola.forms import MateriaForm
 from escola.models import Turma, MateriaDaTurma
-from escola.utils import dar_permissao_perm_a_user_of_level
+
+logger = logging.getLogger(__name__)
 
 
-@permission_required_obj('escola.can_add_materia', (Turma, 'pk', 'turma_pk'))
 def add_materia(request, turma_pk):
-    # FIXME Adicionar permissões, a lista de permissões do grupo LIDER, VICELIDER e REGENTE da turma;
+    turma = get_object_or_404(Turma, pk=turma_pk)
+    if not has_object_permission('add_materia', request.user, turma):
+        return redirect_to_login(request.get_full_path())
     if request.method == 'POST':
         form = MateriaForm(request.POST)
         if form.is_valid():
@@ -27,8 +31,7 @@ def add_materia(request, turma_pk):
             materia.professor = form.cleaned_data['professor']
             materia.abreviacao = form.cleaned_data['abreviacao']
             materia.save()
-            dar_permissao_perm_a_user_of_level('can_edit_materia', 1, get_object_or_404(Turma, pk=turma_pk), materia)
-            dar_permissao_perm_a_user_of_level('can_delete_materia', 2, get_object_or_404(Turma, pk=turma_pk), materia)
+            # TODO: 25/04/2019 por wwwvi: Dar permissão ao criador ou algo parecido.
             return HttpResponseRedirect(reverse('escola:list-materias', args=[turma_pk]))
     else:
         form = MateriaForm()
@@ -46,10 +49,11 @@ def list_materias(request, turma_pk):
     return render(request, 'escola/materia/listMaterias.html', context={'materias': materias, 'turma': turma})
 
 
-@permission_required_obj('escola.can_edit_materia', (MateriaDaTurma, 'pk', 'materia_pk'))
 def edit_materia(request, materia_pk):
     materia = get_object_or_404(MateriaDaTurma, pk=materia_pk)
     turma = materia.turma
+    if not has_object_permission('edit_materia', request.user, materia):
+        return redirect_to_login(request.get_full_path())
     if request.method == 'POST':
         form = MateriaForm(request.POST)
         if form.is_valid():
@@ -68,9 +72,10 @@ def edit_materia(request, materia_pk):
     return render(request, 'escola/materia/formMateria.html', context=context)
 
 
-@permission_required_obj('escola.can_delete_materia', (MateriaDaTurma, 'pk', 'materia_pk'))
 def delete_materia(request, materia_pk):
     materia = get_object_or_404(MateriaDaTurma, pk=materia_pk)
+    if not has_object_permission('edit_materia', request.user, materia):
+        return redirect_to_login(request.get_full_path())
     turma = materia.turma
     materia.delete()
     return HttpResponseRedirect(reverse('escola:list-materias', args=[turma.pk]))
